+++ src/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -17,3 +18,13 @@ root.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
+
+// Register service worker for PWA functionality
+serviceWorkerRegistration.register({
+  onSuccess: (registration) => {
+    console.log('PWA: Service Worker registered successfully');
+  },
+  onUpdate: (registration) => {
+    console.log('PWA: New app version available');
+  }
+});